# crates/full-node/sov-api-spec/openapi-v3.yaml
...
openapi: 3.0.2
info:
  title: Sovereign Full Node API
  version: 0.1.0
  description: |-
    This is the REST API for the Sovereign SDK full node.
  contact:
    name: Sovereign Labs
    url: https://github.com/Sovereign-Labs/sovereign-sdk-wip
    email: info@sovereign.xyz
  license:
    name: Apache-2.0
    url: https://github.com/Sovereign-Labs/sovereign-sdk-wip/blob/nightly/LICENSE.md
servers:
  - url: http://localhost:12346
    description: Localhost
security: []
paths:
...
  /ledger/events:
    get:
      tags:
        - Ledger
      summary: Lists events.
      operationId: list_events
      parameters:
        - $ref: "#/components/parameters/pageType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/pageCursor"
      responses:
        "200":
          description: "Returns a list of events"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEvent"
        "400":
          $ref: "#/components/responses/ApiErrorResponse"
...
  /sequencer/txs:
    post:
      summary: Submit a transaction to the rollup
      description: Submit a transaction to the rollup
      operationId: acceptTx
      tags:
        - Sequencer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                body:
                  type: string
                  format: base64
              required:
                - body
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxInfoWithConfirmation"
        "400":
          "$ref": "#/components/responses/ApiErrorResponse"
        "403":
          "$ref": "#/components/responses/ApiErrorResponse"
        "500":
          "$ref": "#/components/responses/ApiErrorResponse"
        "503":
          "$ref": "#/components/responses/ApiErrorResponse"
  /sequencer/txs/{txHash}/status:
    get:
      summary: Get a transaction status by hash
      operationId: get_tx_status
      tags:
        - Sequencer
      parameters:
        - $ref: "#/components/parameters/txHash"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxInfo"
        "404":
          $ref: "#/components/responses/ApiErrorResponse"
  /sequencer/txs/{txHash}:
    get:
      summary: Get a transaction by hash
      operationId: sequencer_get_tx
      tags:
        - Sequencer
      parameters:
        - $ref: "#/components/parameters/txHash"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiAcceptedTx"
        "404":
          $ref: "#/components/responses/ApiErrorResponse"
  /sequencer/unstable/events/{eventOffset}:
    get:
      summary: |
        Get a soft-confirmed event by event number.

        Note that this endpoint is unstable and it may not always
        work as expected, sometimes an event might be missing
        that was already soft-confirmed.
      operationId: sequencer_get_event
      tags:
        - Sequencer
      parameters:
        - $ref: "#/components/parameters/eventOffset"
      responses:
        "200":
          $ref: "#/components/responses/LedgerEvent"
        "404":
          $ref: "#/components/responses/ApiErrorResponse"
  /sequencer/unstable/events:
    get:
      summary: |
        Get a list of soft-confirmed events.

        Note that this endpoint is unstable and it may not always
        work as expected, sometimes an event might be missing
        that was already soft-confirmed.
      operationId: sequencer_list_events
      parameters:
        - $ref: "#/components/parameters/pageType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/pageCursor"
      tags:
        - Sequencer
      responses:
        "200":
          $ref: "#/components/responses/PaginatedLedgerEvents"
        "404":
          $ref: "#/components/responses/ApiErrorResponse"
...
components:
  parameters:
    pageType:
      name: page
      in: query
      description: Pagination type (first, next, or last)
      schema:
        type: string
        enum: [first, next, last]
        default: first
    pageSize:
      name: page[size]
      in: query
      description: Number of items per page
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 25
    pageCursor:
      name: page[cursor]
      in: query
      description: Cursor for next page (required when page=next)
      schema:
        type: string
...
    eventOffset:
      name: eventOffset
      in: path
      required: true
      schema:
        type: integer
        format: uint64
...
    txHash:
      name: txHash
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/TxHash"
  responses:
...
    LedgerEvent:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LedgerEvent"
...
    PaginatedLedgerEvents:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              next_cursor:
                type: string
              items:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEvent"
            required:
              - items
...
    ApiErrorResponse:
      description: The request sent was invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    TxReceipt:
      type: object
      properties:
        result:
          type: string
          enum:
            - successful
            - reverted
            - skipped
      required:
        - result
...
    LedgerEvent:
      type: object
      properties:
        type:
          type: string
          enum:
            - event
        number:
          type: integer
          format: uint64
        key:
          type: string
        value:
          type: object
          additionalProperties: true
        module:
          $ref: "#/components/schemas/ModuleRef"
        tx_hash:
          $ref: "#/components/schemas/Hash"
      required:
        - type
        - number
        - key
        - value
        - module
    ModuleRef:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
    Hash:
      type: string
      title: Hash
      pattern: ^0x[a-fA-F0-9]{64}$
      description: Hash
...
    TxHash:
      type: string
      pattern: ^0x[a-fA-F0-9]+$
      description: The hash of a transaction
    TxInfo:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/TxHash"
        status:
          $ref: "#/components/schemas/TxStatus"
      required:
        - id
        - status
    TxInfoWithConfirmation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/TxHash"
        status:
          $ref: "#/components/schemas/TxStatus"
        tx_number:
          type: integer
          format: uint64
        events:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEvent"
        receipt:
          $ref: "#/components/schemas/TxReceipt"
      required:
        - id
        - status
    ApiAcceptedTx:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/TxHash"
        tx_number:
          type: integer
          format: uint64
        events:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEvent"
        receipt:
          $ref: "#/components/schemas/TxReceipt"
      required:
        - id
        - tx_number
        - events
        - receipt
    TxStatus:
      type: string
      enum:
        - unknown
        - dropped
        - submitted
        - published
        - processed
        - finalized
...
    ApiError:
      type: object
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code related to this error
        title:
          type: string
        details:
          type: object
      required:
        - status
        - title
        - details
...